Frontend_desktop/requirements.txt
customtkinter
requests
python-dotenv
pillow
sqlalchemy
passlib[bcrypt]